+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## UNRELEASED
 ### Fixed
  - bug in the lexer that was not treating `constructor` as an identifier (#66)
+ - bug when printing diagnostics that would sometimes fail to find the line in question (#68)